finalized basic functions

+++ app/Http/Controllers/UserController.php
@@ -9,6 +9,7 @@ use Spatie\Permission\Models\Role;
 use DB;
 use Hash;
 use Illuminate\Support\Arr;
+use Illuminate\Database\Eloquent\Builder;
 
 
 class UserController extends Controller
@@ -27,10 +28,12 @@ class UserController extends Controller
      */
     public function index(Request $request)
     {
-        $data = User::with('roles')->with('userTasks')->latest()->paginate(5);
-        return view('users.index',compact('data'));
+        $data = User::whereDoesntHave('roles', function (Builder $query) {
+            $query->where('name', 'like', 'Admin');
+        })->with('userTasks')->latest()->paginate(10);
+        return view('users.index', compact('data'));
     }
-    
+
     /**
      * Show the form for creating a new resource.
      *
@@ -38,10 +41,10 @@ class UserController extends Controller
      */
     public function create()
     {
-        $roles = Role::pluck('name','name')->all();
-        return view('users.create',compact('roles'));
+        $roles = Role::pluck('name', 'name')->all();
+        return view('users.create', compact('roles'));
     }
-    
+
     /**
      * Store a newly created resource in storage.
      *
@@ -56,18 +59,18 @@ class UserController extends Controller
             'password' => 'required|same:confirm-password',
             'roles' => 'required'
         ]);
-        
+
         $input = $request->all();
         $input['password'] = Hash::make($input['password']);
-        
+
         $user = User::create($input);
         $user->assignRole($request->input('roles'));
         // activity()->causedby(auth()->user()->id)->performedOn($user)->log('ate ass');
-    
+
         return redirect()->route('users.index')
-                        ->with('success','User created successfully');
+            ->with('success', 'User created successfully');
     }
-    
+
     /**
      * Display the specified resource.
      *
@@ -77,9 +80,9 @@ class UserController extends Controller
     public function show($id)
     {
         $user = User::find($id);
-        return view('users.show',compact('user'));
+        return view('users.show', compact('user'));
     }
-    
+
     /**
      * Show the form for editing the specified resource.
      *
@@ -89,12 +92,12 @@ class UserController extends Controller
     public function edit($id)
     {
         $user = User::find($id);
-        $roles = Role::pluck('name','name')->all();
-        $userRole = $user->roles->pluck('name','name')->all();
-    
-        return view('users.edit',compact('user','roles','userRole'));
+        $roles = Role::pluck('name', 'name')->all();
+        $userRole = $user->roles->pluck('name', 'name')->all();
+
+        return view('users.edit', compact('user', 'roles', 'userRole'));
     }
-    
+
     /**
      * Update the specified resource in storage.
      *
@@ -106,28 +109,28 @@ class UserController extends Controller
     {
         $this->validate($request, [
             'name' => 'required',
-            'email' => 'required|email|unique:users,email,'.$id,
+            'email' => 'required|email|unique:users,email,' . $id,
             'password' => 'same:confirm-password',
             'roles' => 'required'
         ]);
-    
+
         $input = $request->all();
-        if(!empty($input['password'])){ 
+        if (!empty($input['password'])) {
             $input['password'] = Hash::make($input['password']);
-        }else{
-            $input = Arr::except($input,array('password'));    
+        } else {
+            $input = Arr::except($input, array('password'));
         }
-    
+
         $user = User::find($id);
         $user->update($input);
-        DB::table('model_has_roles')->where('model_id',$id)->delete();
-    
+        DB::table('model_has_roles')->where('model_id', $id)->delete();
+
         $user->assignRole($request->input('roles'));
-    
+
         return redirect()->route('users.index')
-                        ->with('success','User updated successfully');
+            ->with('success', 'User updated successfully');
     }
-    
+
     /**
      * Remove the specified resource from storage.
      *
@@ -136,8 +139,12 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        if (User::find($id)->hasRole('Admin')) {
+            return redirect()->route('users.index')
+                ->with('failed', 'Can\'t delete Admins mate');
+        }
         User::find($id)->delete();
         return redirect()->route('users.index')
-                        ->with('success','User deleted successfully');
+            ->with('success', 'User deleted successfully');
     }
 }
